third_party/blink/web_tests/http/tests/security/mixedContent/websocket/insecure-websocket-in-secure-page-allowed.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/testharness.js"></script>
<script src="/js-test-resources/testharnessreport.js"></script>
<script>

if (window.testRunner) {
    testRunner.setPopupBlockingEnabled(false);
    // Allow insecure WebSockets.
    testRunner.overridePreference("WebKitAllowRunningInsecureContent", true);
}

var test = async_test("Test that an insecure WebSocket connection is blocked when created in an HTTPS resource.");
test.step(function ()
{
    window.addEventListener("message", test.step_func(function (evt)
    {
        assert_equals(evt.data, "DONE", "evt.data");
        test.done();
    }), false);

    window.open("https://127.0.0.1:8443/security/mixedContent/websocket/resources/expect-successful-construction.html");
});

</script>
</head>
<body>
</body>
</html>
